DanielBalducciReadingone
Into time .com by RAFAËL ROZENDAAL was interesting to me because I never truly communicated with a website that was dedicated to art.
Into time .com explores the browsers potential to create different but similar effects by just the click of a mouse. 
Exploring his other sites like hybrid moment and on and off .org gave me insight to how web based art can be so simple yet interactive. 
I found it interesting how all of Rafael’s net based art have different ways of interacting. 
For example, one of his works utilizes the placement of the cursor while the other utilizes the click function of the mouse. 
What I liked about into time .com was that you can make a lot of different images by just clicking what was on the website. 
The ease of simplicity and potential variances is successful for user interaction.

Nine eyes of google street view by Jon Rafman was another piece I found interesting because of the photography captured by a machine that is without human bias.
The way in which we see the world cannot be more honest through the lens of the google street view, and it captures some moments that are questionable or interesting to see. 
My favorite example of this are the more disturbing images such as a women dragging another women by her hair in the middle of the street. 
The scenes displayed in https://webrecorder.io/ shows some remarkable moments or just ironic glitched images of a human or landscape. 
Most images covered by this automated photographic machine tends to hold a creepy vibe to most its pictures as they look uninterrupted. 
I myself found this interesting because this past summer, My step dad found himself within the google street view along side with my mother somewhere in the westside of Japan. 
Although their faces are blurred, you can tell that my parents identified the vehicle and its purpose and somewhat posed for the camera.

Last but not least, my favorite of the 3 that I’ve seen might be “the revolving internet”  by CONSTANT DULLAART.
It is an experience that is strange but familiar enough and the fact that web functionality is a part of the piece speaks to me. 
What I found interesting is the part where you’re not in control of the rotation and how the name of the website is literally “the revolving internet.” 
The fusion makes sense and to have functionality of the scrolling of the site as well as the primary search function is a weird touch that also brings to question to my mind “How is this art.” 
Perhaps the best feature is its simplicity among the things that are so familiar to me in everyday internet browsing.