Update XiaoWuReadingOne
The first work that interested me is The Dollie Clone Series created by Lynn Hershman Leeson. Basically, Lynn replaced a doll’s eye with a webcam and make it a ‘cyborg’. The webcam was connected to the internet and allows the audience seeing things from the doll’s sight. In addition, the audience are able to control the direction of the webcam by remote manipulate on a website. I love this work because it makes the doll out of the ordinary. It gives a creepy feeling made by the contemporary technology. I also love how this work let audiences participate into it. In my opinion, the participation extended the audience’s perspective, and make the doll an actual cyborg -- part life and part machine. 

Antoni Muntadas’ The File Room is another interesting work. It has less digital element but more related to politics. According to the Net Art Anthology, it was a temporary physical installation about deleted and suppressed files, but lately kept and saved by other organization. In my opinion, it is a satirical art which constituted by people around the world. All data base were submitted by participator while the project itself didn’t participate the evaluation. Therefore, I personally believe the facticity is questionable, while the very important thing of this work is the existence.