Add files via upload
Heres the only documentation I have from the webpage. I used code from A-Frame to make the website but I couldn't figure it out. The idea was to create a three dimensional social network where people can knowingly be catfish and express alter egos via web. I was inspired by another Artist Brad Phillips who does similar things but outside of the internet. I wanted to create a platform where people could act out their fantasy of being a curator for the Guggenheim or be an astronaut for NASA. Brad makes fake business cards with a Logo from book publishers and his name on it with a made up title and I wanted to extend that through the internet.